refactor(sender): 弹窗文件归入 dialogs/ 目录
- dialogs.py → dialogs/pre_send.py
- task_builder_dialog.py → dialogs/task_builder.py
- task_detail_dialog.py → dialogs/task_detail.py
- 更新 page.py 导入路径
