+++ src/danmaku_sender/ui/sender/components/dialogs/task_detail.py
